Added Error handling in: BranchC, EmpC, ItemC, PatientC, VisitC. And introduced exception folder

+++ backend/src/main/java/com/jwt/spring_security/repo/ItemRepo.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ItemRepo extends JpaRepository<Item, Integer> {
+public interface ItemRepo extends JpaRepository<Item, Long> {
 
     Item findByItemID(Long id);
